+++ testing/__init__.py
@@ -8,8 +8,8 @@ __all__ = [
     "HAS_PYTEST",
     "NO_ONE_CHANGE_VERSION",
 ]
-__version__ = "1.2.0"
-__version_tuple__ = (1, 2, 0)
+__version__ = "1.2.1"
+__version_tuple__ = (1, 2, 1)
 
 try:
     import pytest